implement lock-free control of session parameter "bpm"
this feature required a re-work of the way frame counting is handled.
before, each sequence kept its own frame count, which required each
sequence to also keep track of the frames-per-tick, which is dependent
on the BPM (and the ticks-per-step). now, the *session* keeps track of
frame count, calling gs_sequence_tick() every time a tick boundary is
crossed.