sequence: implement buf_off to store "off" events separate from the microgrid
buf_off is a circular buffer of midi_packet's (nticks of them) which is
always running forward in time, regardless of the state of the main
playhead. it is dynamically populated with a note-off each time a
note-on fires from the callback thread. when a note-off fires, it is
auto-removed from the buf_off.