introduce branching setter methods for control parameters
this commit switches over to setter methods (e.g. sequence_set_trig())
which branch between either an immediate setter which executes within
the calling thread (set_trig_now()), or a lock-free message queue which
is served within the audio thread. this is needed in order to prevent
the message queue from overflowing while e.g. setting trigs in a non-
playing sequence before adding it to the session.

for both the session and the sequence classes, the branching is
determined by a flag (is_playing) which is set and accessed from the
UI thread only.