re-implement note on and control change (CC)
This commit introduces a sorting algorithm for handling the possibility
that MIDI events may be registered out-of-order in the JACK buffer,
resulting in the following error (and seg-fault):

jack_midi_event_reserve: time parameter is earlier than last reserved event

This was not a problem with the previous "microgrid" version, because
all events were discretized to a tick, so at worst, they end up on the
same frame (which is allowed by the JACK API). Now, we have microtiming
down to the frame resolution.

This commit leaves out the re-implementation of note off.