implement note-off buffer as an array of linked lists
This fixes the issue of hanging notes caused by note-off's overwriting
each other. Previously, the note-off buffer had only one slot per
frame, so if a later note's off-message lands on the same frame as that
of a previous note, the latter gets overwritten. Now, the buffer is
implemented as an array of linked lists, so each frame can point to an
arbritrarily long (up a limit set by the offHeat_t instance) set of off
messages.